requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
avalanche-lib==0.3.1
azure-common==1.1.28
azure-core==1.29.3
azure-identity==1.14.0
azure-mgmt-compute==30.1.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
boto3==1.26.133
botocore==1.29.133
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.27.0
cryptography==41.0.3
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.6
diversipy==0.8
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
EoN==1.1
executing==0.8.3
fastjsonschema==2.17.1
filelock==3.9.0
flatbuffers==22.9.24
fonttools==4.33.3
forbiddenfruit==0.1.4
fpzip==1.2.0
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.3
gast==0.4.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.46.3
h5py==3.7.0
huggingface-hub==0.14.1
hydra-core==1.2.0
idna==3.3
imageio==2.22.1
import-ipynb==0.1.4
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipykernel==6.13.0
ipython==8.3.0
ipywidgets==8.0.4
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonref==1.1.0
jsonschema==4.17.3
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-widgets==3.0.5
kaggle==1.5.12
keopscore==2.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kmodes==0.12.2
libclang==14.0.6
lightly==1.2.19
lightly-utils==0.0.2
lit==16.0.6
llvmlite==0.39.1
Markdown==3.3.7
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
monotonic==1.6
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.0.5
multidict==6.0.2
multiprocess==0.70.14
mypy-extensions==0.4.3
nbformat==5.9.0
neptune==1.2.0
nest-asyncio==1.5.5
networkx==3.1
numba==0.56.4
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow-SIMD==9.0.0.post1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
portalocker==2.7.0
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.12.0
PyJWT==2.7.0
pykeops==2.1.2
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pysql-framework==1.0.4
python-dateutil==2.8.2
python-slugify==6.1.2
pytorch-lightning==1.6.4
pytorchcv==0.0.67
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.0.0
quadprog==0.1.11
regex==2023.5.5
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
s3transfer==0.6.1
scikit-image==0.19.3
scikit-learn==1.3.1
scipy==1.8.1
seaborn==0.12.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sentry-sdk==1.13.0
setproctitle==1.3.2
simplejson==3.19.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
# Editable Git install with no remote (src==0.0.0)
-e /home/qihang/repos/network_topology
stack-data==0.2.0
style==1.1.0
swagger-spec-validator==3.0.3
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.8.12
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.7.1
update==0.0.1
urllib3==1.26.16
wandb==0.13.9
wcwidth==0.2.5
websocket-client==1.5.1
Werkzeug==2.1.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xxhash==3.2.0
yarl==1.7.2
zipp==3.8.0
